refactor(LottoResult): winningAmount 상태값 제거
1. LottoResult에서 winningAmount 상태값 제거
2. winningAmount(당첨금총액) 계산 메서드 추가
3. 관련 프로덕션/테스트 코드 수정

+++ src/main/java/step3/domain/Ranking.java
@@ -13,7 +13,7 @@ public enum Ranking {
     FIRST(6, 2_000_000_000);
 
     private final int hitCount;
-    private final int winningMoney;
+    private final long winningMoney;
 
     Ranking(int hitCount, int winningMoney) {
         this.hitCount = hitCount;
@@ -31,7 +31,7 @@ public enum Ranking {
         return hitCount;
     }
 
-    public double getWinningMoney() {
+    public long getWinningMoney() {
         return winningMoney;
     }
